move all tests from test_validation_opt into test_validation
